+++ examples/custom_checkers/e9989_pycodestyle/e101_no_error.py
@@ -1,0 +1,4 @@
+if True:
+    # Indentation contains only tab character
+    # OR Indentation contains only space characters
+    pass