+++ examples/custom_checkers/e9989_pycodestyle/e204_error.py
@@ -1,0 +1,8 @@
+@ property  # one whitespace after decorator
+def func1():
+    return 'hello world'
+
+
+@   property  # multiple whitespaces after decorator
+def func2():
+    return 'hello world'