+++ python_ta/reporters/node_printers.py
@@ -263,6 +263,21 @@ def render_pep8_errors_e201_e202_e203_e211(msg, _node, source_lines=None):
     yield from render_context(line + 1, line + 3, source_lines)
 
 
+def render_pep8_errors_e204(msg, _node, source_lines=None):
+    """Render a PEP8 whitespace after decorator '@' message"""
+    line = msg.line
+    res = re.search(r"column (\d+)", msg.msg)
+    col = int(res.group().split()[-1])
+    # calculates the length of the leading whitespaces by subtracting the length of everything after the first character after stripping all leading whitespaces from the total line length
+    curr_idx = (
+        col + len(source_lines[line - 1][col:]) - len(source_lines[line - 1][col + 1 :].lstrip())
+    )
+
+    yield from render_context(line - 2, line, source_lines)
+    yield (line, slice(col, curr_idx), LineType.ERROR, source_lines[line - 1])
+    yield from render_context(line + 1, line + 3, source_lines)
+
+
 def render_pep8_errors_e221(msg, _node, source_lines=None):
     """Render a PEP8 multiple spaces before operator message."""
     line = msg.line
@@ -287,8 +302,8 @@ def render_pep8_errors_e222(msg, _node, source_lines=None):
     yield from render_context(line + 1, line + 3, source_lines)
 
 
-def render_pep8_errors_e223(msg, _node, source_lines=None):
-    """Render a PEP8 tab before operator message."""
+def render_pep8_errors_e223_and_e274(msg, _node, source_lines=None):
+    """Render a PEP8 tab before operator message and a PEP8 tab before keyword message."""
     line = msg.line
     res = re.search(r"column (\d+)", msg.msg)
     col = int(res.group().split()[-1])
@@ -315,12 +330,53 @@ def render_pep8_errors_e224_and_e273(msg, _node, source_lines):
     yield from render_context(line + 1, line + 3, source_lines)
 
 
+def render_pep8_errors_e225(msg, _node, source_lines):
+    """Render a PEP8 missing whitespace around operator message"""
+    line = msg.line
+    res = re.search(r"column (\d+)", msg.msg)
+    col = int(res.group().split()[-1])
+    curr_idx = col + 1
+
+    two_char_operators = {
+        "==",
+        ">=",
+        "<=",
+        "!=",
+        ":=",
+        "&=",
+        "->",
+        "%=",
+        "/=",
+        "+=",
+        "-=",
+        "*=",
+        "|=",
+        "^=",
+        "@=",
+    }
+    three_char_operators = {"//=", ">>=", "<<=", "**="}
+    # highlight multiple characters for operators that are longer than one character
+    if source_lines[line - 1][col : col + 2] in two_char_operators:
+        curr_idx += 1
+    elif source_lines[line - 1][col : col + 3] in three_char_operators:
+        curr_idx += 2
+
+    yield from render_context(line - 2, line, source_lines)
+    yield (line, slice(col, curr_idx), LineType.ERROR, source_lines[line - 1])
+    yield from render_context(line + 1, line + 3, source_lines)
+
+
 def render_pep8_errors_e226(msg, _node, source_lines):
     """Render a PEP8 missing whitespace around arithmetic operator message"""
     line = msg.line
     res = re.search(r"column (\d+)", msg.msg)
     col = int(res.group().split()[-1])
     end_idx = col + 1
+
+    multi_char_operators = {"//"}
+    # highlight multiple characters for arithmetic operators that are longer than one character
+    if source_lines[line - 1][col : col + 2] in multi_char_operators:
+        end_idx += 1
 
     yield from render_context(line - 2, line, source_lines)
     yield (line, slice(col, end_idx), LineType.ERROR, source_lines[line - 1])
@@ -357,6 +413,17 @@ def render_pep8_errors_e228(msg, _node, source_lines=None):
         LineType.ERROR,
         source_lines[line - 1] + "  # INSERT A SPACE BEFORE AND AFTER THE % OPERATOR",
     )
+    yield from render_context(line + 1, line + 3, source_lines)
+
+
+def render_pep8_errors_e231(msg, _node, source_lines=None):
+    line = msg.line
+    res = re.search(r"column (\d+)", msg.msg)
+    col = int(res.group().split()[-1])
+    curr_idx = col + 1
+
+    yield from render_context(line - 2, line, source_lines)
+    yield (line, slice(col, curr_idx), LineType.ERROR, source_lines[line - 1])
     yield from render_context(line + 1, line + 3, source_lines)
 
 
@@ -433,6 +500,18 @@ def render_pep8_errors_e266(msg, _node, source_lines=None):
         LineType.ERROR,
         source_lines[line - 1] + "  # THERE SHOULD ONLY BE ONE '#'",
     )
+    yield from render_context(line + 1, line + 3, source_lines)
+
+
+def render_pep8_errors_e271(msg, _node, source_lines=None):
+    """Render a PEP8 multiple spaces after keyword message."""
+    line = msg.line
+    res = re.search(r"column (\d+)", msg.msg)
+    col = int(res.group().split()[-1])
+    curr_idx = col + len(source_lines[line - 1][col:]) - len(source_lines[line - 1][col:].lstrip())
+
+    yield from render_context(line - 2, line, source_lines)
+    yield (line, slice(col, curr_idx), LineType.ERROR, source_lines[line - 1])
     yield from render_context(line + 1, line + 3, source_lines)
 
 
@@ -570,6 +649,17 @@ def render_pep8_errors_e306(msg, _node, source_lines=None):
     yield from render_context(msg.line, msg.line + 2, source_lines)
 
 
+def render_pep8_errors_e502(msg, _node, source_lines=None):
+    """Render a PEP8 the backslash is redundant between brackets."""
+    line = msg.line
+    res = re.search(r"column (\d+)", msg.msg)
+    col = int(res.group().split()[-1])
+
+    yield from render_context(line - 2, line, source_lines)
+    yield (line, slice(col, col + 1), LineType.ERROR, source_lines[line - 1])
+    yield from render_context(line + 1, line + 3, source_lines)
+
+
 def render_missing_return_statement(msg, node, source_lines=None):
     """
     Render a missing return statements message
@@ -661,11 +751,16 @@ RENDERERS = {
     "E201": render_pep8_errors_e201_e202_e203_e211,
     "E202": render_pep8_errors_e201_e202_e203_e211,
     "E203": render_pep8_errors_e201_e202_e203_e211,
+    "E204": render_pep8_errors_e204,
+    "E211": render_pep8_errors_e201_e202_e203_e211,
     "E221": render_pep8_errors_e221,
     "E222": render_pep8_errors_e222,
-    "E223": render_pep8_errors_e223,
+    "E223": render_pep8_errors_e223_and_e274,
     "E224": render_pep8_errors_e224_and_e273,
+    "E225": render_pep8_errors_e225,
+    "E231": render_pep8_errors_e231,
     "E273": render_pep8_errors_e224_and_e273,
+    "E274": render_pep8_errors_e223_and_e274,
     "E226": render_pep8_errors_e226,
     "E227": render_pep8_errors_e227,
     "E228": render_pep8_errors_e228,
@@ -674,6 +769,7 @@ RENDERERS = {
     "E262": render_pep8_errors_e262,
     "E265": render_pep8_errors_e265,
     "E266": render_pep8_errors_e266,
+    "E271": render_pep8_errors_e271,
     "E272": render_pep8_errors_e272,
     "E275": render_pep8_errors_e275,
     "E301": render_pep8_errors_e301,
@@ -682,6 +778,7 @@ RENDERERS = {
     "E304": render_pep8_errors_e304,
     "E305": render_pep8_errors_e305,
     "E306": render_pep8_errors_e306,
+    "E502": render_pep8_errors_e502,
 }